Cleanup and move test-only code into the test utils library
